Add support for IPv6 to proxy from create_connection.
This change adds support for making IPv6 connections to the proxy by
changing create_connection to use getaddrinfo. This fully supports IPv6
when a hostname only resolves to it, as well as ensuring that the socket
family and type are automatically set correctly.

The bulk of this code was drawn from the urllib3 equivalent of this
method. This is compatible with the license here.

This problem does not exist when using socksocket objects directly, as
they can have their type set explicitly: it's only a problem when using
create_connection.

Note that this commit *also* strips square brackets from the remote host
if it is an IPv6 address. This is useless right now as there is no
support in PySocks for contacting raw IPv6 addresses, but I intend to
add that support as well.