Timeout earlier in connection.
It is possible when setting timeouts on socksocket objects to find that
they hang indefinitely during connection or negotiation. This is a
less-than-ideal user experience. This change moves the socket timeout
setting earlier in the connection setup so that it applies to the actual
TCP connection and the negotiation, if there is any.